fix: wrong collect logic when thread stuck
Previously we always try to scan the stack precisely if the
thread is stuck even when conservative stack scan feature
is enabled. This may cause unexpected behaviours when using
stucking apis.

This commit also fixed a bug in optimization pipeline in O3 level.
Previously we are missing some simplification pass on this level,
preventing opt like tail call elimination in some cases.

The last bug fixed in this commiy is related to escape analysis.
The inferenced tuple types are always atomic due to a typo, which
causing invalid memory access error when dealing with non atomic
tuples.