fix: windows gc scan invalid memory access  exception
Stack memory is not automatically initialized when the new function was called,
so it's important to tolerate the malformed data during conservative stack scanning.

Before this commit, the exception may happen in following situation:

Fn A first called fn B, which allocated a pointer `P1`, it's a complex type and was
referenced by a stck slot in B's stack frame. During execution of B, it assign `P1`
to a live data created by A (which means it now was both referenced by B's stack slot
and A's data). After B was done, A triggered a GC, which evacuates `P1` to `P2`,
and correct all live references of `P1` to `P2`, which doesn't include the old stack slot of `B`
as B is already done execution. Then A called C, which stack frame share most space of the old
frame of `B`, and as the stack space will not be cleaned after used, there's a reference of
`P1` on it! Then when C trigger a GC again, the evacuated 'dead' `P1` will be treated as a
live complex type object, which lead to invalid memory access issue when calling its mark fn
in gc rtti field.